Fix How claims work button not displaying
