+++ Projects/Home/Sources/Viewables/ActiveSection.swift
@@ -45,15 +45,9 @@ extension ActiveSection: Presentable {
         }
 
         bag += section.appendSpacingAndDumpOnDispose(.inbetween)
-
-        let howClaimsWorkButton = Button(
-            title: L10n.ClaimsExplainer.title,
-            type: .iconTransparent(
-                textColor: .brand(.primaryTintColor),
-                icon: .left(image: hCoreUIAssets.infoSmall.image, width: .smallIconWidth)
-            )
-        )
-        bag += section.append(howClaimsWorkButton.alignedTo(alignment: .center))
+        
+        let howClaimsWorkButton = HowClaimsWorkButton()
+        bag += section.append(howClaimsWorkButton)
 
         bag += store.stateSignal.atOnce().map { ($0.claims?.count ?? 0) > 0 }
             .onValue { hasClaims in
@@ -70,36 +64,6 @@ extension ActiveSection: Presentable {
                         backgroundColor: .brand(.secondaryButtonBackgroundColor),
                         textColor: .brand(.secondaryButtonTextColor)
                     )
-            }
-
-        bag += howClaimsWorkButton.onTapSignal.compactMap { section.viewController }
-            .onValue { viewController in
-                var pager = Pager(
-                    title: L10n.ClaimsExplainer.title,
-                    buttonContinueTitle: L10n.ClaimsExplainer.buttonNext,
-                    buttonDoneTitle: L10n.ClaimsExplainer.buttonStartClaim,
-                    pages: []
-                ) { viewController in
-                    store.send(.openClaims)
-                    return Future(.forever)
-                }
-                viewController.present(pager)
-
-                client.fetch(
-                    query: GraphQL.HowClaimsWorkQuery(
-                        locale: Localization.Locale.currentLocale.asGraphQLLocale()
-                    )
-                )
-                .onValue { data in
-                    pager.pages = data.howClaimsWork.map {
-                        ContentIconPagerItem(
-                            title: nil,
-                            paragraph: $0.body,
-                            icon: $0.illustration.fragments.iconFragment
-                        )
-                        .pagerItem
-                    }
-                }
             }
 
         bag += section.append(ConnectPaymentCard())
