+++ Projects/Home/Sources/Viewables/TerminatedSection.swift
@@ -48,6 +48,11 @@ extension TerminatedSection: Presentable {
         bag += claimButton.onTapSignal.onValue {
             store.send(.openClaims)
         }
+        
+        bag += section.appendSpacingAndDumpOnDispose(.inbetween)
+        
+        let howClaimsWorkButton = HowClaimsWorkButton()
+        bag += section.append(howClaimsWorkButton)
 
         return (section, bag)
     }
